fix(minjae/error): 거래완료, 댓글 답글 issue fix #75
