+++ app/components/Comment/SingleComment.js
@@ -24,6 +24,7 @@ const SingleComment = ({ comment, categoryName, num }) => {
     num,
     groupNum: comment.groupNum,
   });
+
   const [updateFormValue, setUpdateFormValue] = useState({
     content: comment.content,
     studentId: auth.id,
@@ -57,6 +58,7 @@ const SingleComment = ({ comment, categoryName, num }) => {
   };
 
   const onOpenUpdate = () => {
+    console.log(updateFormValue);
     setOpenUpdate(!openUpdate);
     if (openReply) {
       setOpenReply(false);
@@ -106,7 +108,7 @@ const SingleComment = ({ comment, categoryName, num }) => {
     }
   };
 
-  const onUpdate = (e) => {
+  const onUpdate = async (e) => {
     e.preventDefault();
 
     const { content, categoryName, num, groupNum, commentNum } =
@@ -124,7 +126,7 @@ const SingleComment = ({ comment, categoryName, num }) => {
     if (body.content.length === 0) {
       alert("댓글에 수정 할 내용을 입력해주세요.");
     } else {
-      dispatch({
+      await dispatch({
         type: COMMENT_UPDATE_REQUEST,
         payload: body,
       });
@@ -236,7 +238,7 @@ const SingleComment = ({ comment, categoryName, num }) => {
                   className="comment-content-area update"
                   onChange={onUpdateChange}
                   placeholder="Comment"
-                  defaultValue={comment.content}
+                  defaultValue={comment.content.replace(/[<]br [/][>]/gi, "\n")}
                 />
 
                 <button
